requirements.txt
streamlit
transformers
torch
pandas
python-dotenv
requests
google-generativeai
cohere
openpyxl
